move transfer cmd tests to own module
